trying to avoid some code duplication
see #8
