+++ instaseis/server/routes/seismograms.py
@@ -8,7 +8,6 @@
     (http://www.gnu.org/copyleft/lgpl.html)
 """
 import io
-import re
 import zipfile
 
 import numpy as np
@@ -18,12 +17,9 @@ import tornado.web
 
 from ... import Source, ForceSource, Receiver
 from ...base_instaseis_db import _get_seismogram_times
-from ..util import run_async
+from ..util import run_async, IOQueue, _validtimesetting
 from ..instaseis_request import InstaseisRequestHandler
 from ...helpers import geocentric_to_wgs84_latitude
-
-# Valid phase offset pattern including capture groups.
-PHASE_OFFSET_PATTERN = re.compile(r"(^[A-Za-z0-9^]+)([\+-])([\deE\.\-\+]+$)")
 
 
 @run_async
@@ -124,36 +120,6 @@ def _get_seismogram(db, source, receiver, components, units, dt, kernelwidth,
         callback(byte_strings)
 
 
-class IOQueue(object):
-    """
-    Object passed to the zipfile constructor which acts as a file-like object.
-
-    Iterating over the object yields the data pieces written to it since it
-    has last been iterated over DELETING those pieces at the end of each
-    loop. This enables the server to send unbounded zipfiles without running
-    into memory issues.
-    """
-    def __init__(self):
-        self.count = 0
-        self.data = []
-
-    def flush(self):
-        pass
-
-    def tell(self):
-        return self.count
-
-    def write(self, data):
-        self.data.append(data)
-        self.count += len(data)
-
-    def __iter__(self):
-        for _i in self.data:
-            yield _i
-        self.data = []
-        raise StopIteration
-
-
 def _tolist(value, count):
     value = [float(i) for i in value.split(",")]
     if len(value) not in count:
@@ -171,33 +137,6 @@ def _doublecouple(value):
 
 def _forcesource(value):
     return _tolist(value, (3,))
-
-
-def _validtimesetting(value):
-    try:
-        return obspy.UTCDateTime(value)
-    except:
-        pass
-
-    try:
-        return float(value)
-    except:
-        pass
-
-    m = PHASE_OFFSET_PATTERN.match(value)
-    if m is None:
-        raise ValueError
-
-    operator = m.group(2)
-    if operator == "+":
-        offset = float(m.group(3))
-    else:
-        offset = -float(m.group(3))
-
-    return {
-        "phase": m.group(1),
-        "offset": offset
-    }
 
 
 class SeismogramsHandler(InstaseisRequestHandler):
